Add check for current go version before upgrade in ci script

+++ scripts/ci.sh
@@ -1,23 +1,30 @@
 # This script contains commands to be executed by the CI tool.
 NPROC=$(nproc)
 
+function version_gt() { test "$(printf '%s\n' "$@" | sort -V | head -n 1)" != "$1"; }
+
 update_go() {
-	echo "Update go version..."
-	sudo rm -rf /usr/local/go
-	wget https://dl.google.com/go/go1.13.linux-amd64.tar.gz
-	sudo mkdir /usr/local/golang/1.13 && sudo tar -C /usr/local/golang/1.13 -xzf go1.13.linux-amd64.tar.gz
-	rm go1.13.linux-amd64.tar.gz
+	CURRENT_GO_VERSION=`go version | sed 's/[^0-9.]*\([0-9.]*\).*/\1/'`
+	NEW_GO_VERSION=1.13
+	if version_gt $NEW_GO_VERSION $CURRENT_GO_VERSION; then	
+		echo "Update go version from $CURRENT_GO_VERSION to $NEW_GO_VERSION ..."
+		sudo rm -rf /usr/local/go
+		wget https://dl.google.com/go/go$NEW_GO_VERSION.linux-amd64.tar.gz
+		sudo mkdir /usr/local/golang/$NEW_GO_VERSION && sudo tar -C /usr/local/golang/$NEW_GO_VERSION -xzf go$NEW_GO_VERSION.linux-amd64.tar.gz
+		rm go$NEW_GO_VERSION.linux-amd64.tar.gz
 
-	# remove other Go version from path
-	export PATH=`echo $PATH | sed -e 's|:/usr/local/golang/[1-9.]*/go/bin||'`
+		# remove other Go version from path
+		export PATH=`echo $PATH | sed -e 's|:/usr/local/golang/[1-9.]*/go/bin||'`
 
-	sudo ln -fs /usr/local/golang/1.13/go/bin/go /usr/local/bin/go
+		sudo ln -fs /usr/local/golang/$NEW_GO_VERSION/go/bin/go /usr/local/bin/go
 
-	# setup GOROOT
-	export GOROOT="/usr/local/golang/1.13/go"
+		# setup GOROOT
+		export GOROOT="/usr/local/golang/$NEW_GO_VERSION/go"
 
-	# add new go installation to PATH
-	export PATH="$PATH:/usr/local/golang/1.13/go/bin"
+		# add new go installation to PATH
+		export PATH="$PATH:/usr/local/golang/$NEW_GO_VERSION/go/bin"
+	fi
+	go version
 }
 
 setup_protoc() {
